Fixed flows to throw errors on unsupported envs
A library must not hide anything from the user. If we don't write a
"final" flow handler (such as `else`), there is no way for the user to
know why the call to `Push.create()` or `Push.close` failed. He will
just get nothing and will have to debug his entire application in order
to find out that the library doesn't support his environment.