Handle deprecated requestPermission APi in Safari 12+
The assumption was that Safari would not remove the webkit prefix
without also conforming to the promise-based requestPermission API. They
dropped the prefix and did not update to Promises. This checks the
return value of the requestPermission call and treats it as a Promise
when possible. Otherwise, it always passes the callback version for the
old API.

Ref: https://developer.mozilla.org/en-US/docs/Web/API/Notification/requestPermission

Ref: https://developer.apple.com/library/archive/documentation/NetworkingInternet/Conceptual/NotificationProgrammingGuideForWebsites/LocalNotifications/LocalNotifications.html#//apple_ref/doc/uid/TP40012932-SW1